Check count store exists before deletion

+++ community/record-storage-engine/src/main/java/org/neo4j/internal/batchimport/store/BatchingNeoStores.java
@@ -62,6 +62,7 @@ import org.neo4j.kernel.impl.store.RelationshipGroupStore;
 import org.neo4j.kernel.impl.store.RelationshipStore;
 import org.neo4j.kernel.impl.store.StoreFactory;
 import org.neo4j.kernel.impl.store.StoreType;
+import org.neo4j.kernel.impl.store.cursor.CachedStoreCursors;
 import org.neo4j.kernel.impl.store.format.RecordFormats;
 import org.neo4j.kernel.impl.store.format.RecordStorageCapability;
 import org.neo4j.kernel.impl.store.record.RelationshipGroupRecord;
@@ -69,7 +70,6 @@ import org.neo4j.logging.LogProvider;
 import org.neo4j.logging.internal.LogService;
 import org.neo4j.memory.MemoryTracker;
 import org.neo4j.scheduler.JobScheduler;
-import org.neo4j.kernel.impl.store.cursor.CachedStoreCursors;
 import org.neo4j.storageengine.api.cursor.StoreCursors;
 import org.neo4j.token.DelegatingTokenHolder;
 import org.neo4j.token.TokenHolders;
@@ -206,7 +206,10 @@ public class BatchingNeoStores implements AutoCloseable, MemoryStatsVisitor.Visi
 
     private void deleteCountsStore() throws IOException
     {
-        fileSystem.deleteFile( databaseLayout.countStore() );
+        if ( fileSystem.fileExists( databaseLayout.countStore() ) )
+        {
+            fileSystem.deleteFile( databaseLayout.countStore() );
+        }
     }
 
     public void assertDatabaseIsEmptyOrNonExistent()
